# openapi.yaml
openapi: 3.0.3
info:
  title: Band API
  version: "1.0.0"
  description: |
    Read-only JSON API for festival line-ups, stages, bands and time slots.

    Notes / quirks reflected by the current implementation:
    - `{festivalSlug}` is the festival **slug** (e.g. `dong-open-air-2026`); `{stageSlug}` is the stage **slug** (e.g. `hauptbuhne`); `{bandSlug}` is the band **slug** (e.g. `subway-to-sally`). Slugs are auto-generated from the name (lower-case ASCII) and returned by the list endpoints.
    - `logo` / `image` are returned as absolute URLs built as `{host}/images/band/{logos|images}/{filename}`. When the underlying value is null the URL ends with a trailing slash.
    - `GET /api/festivals`, `.../stages` and `.../stages/{stageSlug}` return compact arrays. `startDate` / `endDate` are `YYYY-MM-DD` strings.
    - On the time-slot and band-detail endpoints, `startTime` / `endTime` are still serialized by Symfony's ObjectNormalizer as a **verbose PHP `DateTime` object** (`$ref: PhpDateTime`). `GET .../upnext` returns a preformatted string.
    - `description` is a **localized object** (`$ref: LocalizedText`) with a `de` (German, the default language) and an `en` (English) field. Either field may be `null`. Clients should display `de` by default and fall back to it when the user's locale is English but `en` is `null` (or vice versa).
servers:
  - url: https://bands.baphomet.club
    description: Production
  - url: http://127.0.0.1:8000
    description: Symfony local dev server

paths:
  /api/festivals:
    get:
      operationId: listFestivals
      summary: List all festivals
      responses:
        "200":
          description: Array of festivals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Festival"

  /api/festivals/{festivalSlug}/bands:
    get:
      operationId: listFestivalBands
      summary: List the bands playing a festival
      parameters:
        - $ref: "#/components/parameters/FestivalSlug"
      responses:
        "200":
          description: Array of bands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Band"

  /api/festivals/{festivalSlug}/stages:
    get:
      operationId: listFestivalStages
      summary: List the stages of a festival
      parameters:
        - $ref: "#/components/parameters/FestivalSlug"
      responses:
        "200":
          description: Array of stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stage"

  /api/festivals/{festivalSlug}/stages/{stageSlug}:
    get:
      operationId: getFestivalStageInfo
      summary: Get a stage of a festival by slug
      parameters:
        - $ref: "#/components/parameters/FestivalSlug"
        - $ref: "#/components/parameters/StageSlug"
      responses:
        "200":
          description: Array of matching stages (usually one)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stage"

  /api/festivals/{festivalSlug}/stages/{stageSlug}/timeslots:
    get:
      operationId: listStageTimeSlots
      summary: List the time slots of a stage, ordered by start time
      parameters:
        - $ref: "#/components/parameters/FestivalSlug"
        - $ref: "#/components/parameters/StageSlug"
      responses:
        "200":
          description: Array of time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

  /api/festivals/{festivalSlug}/stages/{stageSlug}/upnext:
    get:
      operationId: getStageUpNext
      summary: Get the next time slot (globally, by current server time)
      description: |
        Returns the next upcoming time slot based on the server's current time (Europe/Berlin).
        Note: the current implementation ignores the `{festivalSlug}` and `{stageSlug}` parameters
        and resolves the next slot globally.
      parameters:
        - $ref: "#/components/parameters/FestivalSlug"
        - $ref: "#/components/parameters/StageSlug"
      responses:
        "200":
          description: Single-element array with the next slot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UpNext"

  /api/festivals/{festivalSlug}/{bandSlug}:
    get:
      operationId: getFestivalBand
      summary: Get a band's detail and its slot at a festival
      description: |
        Returns the band plus its start/end time and stage for the given festival.
        Catch-all route — declared with lower priority than the literal `/bands` and `/stages` routes.
      parameters:
        - $ref: "#/components/parameters/FestivalSlug"
        - name: bandSlug
          in: path
          required: true
          description: Band slug
          schema: { type: string }
          example: "subway-to-sally"
      responses:
        "200":
          description: Band detail with slot info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandDetail"

components:
  parameters:
    FestivalSlug:
      name: festivalSlug
      in: path
      required: true
      description: Festival slug, e.g. `dong-open-air-2026`
      schema: { type: string }
      example: "dong-open-air-2026"
    StageSlug:
      name: stageSlug
      in: path
      required: true
      description: Stage slug, e.g. `hauptbuhne`
      schema: { type: string }
      example: "hauptbuhne"

  schemas:
    Festival:
      type: object
      description: Returned by `GET /api/festivals`.
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: "Dong Open Air 2026" }
        slug: { type: string, example: "dong-open-air-2026" }
        startDate: { type: string, format: date, example: "2026-07-16" }
        endDate: { type: string, format: date, example: "2026-07-18" }

    Band:
      type: object
      description: Band as returned by `GET /api/festivals/{festivalSlug}/bands`.
      properties:
        id: { type: integer, example: 8 }
        name: { type: string, example: "Amorphis" }
        slug: { type: string, example: "amorphis" }
        genre: { type: string, nullable: true, example: null }
        logo:
          type: string
          description: Absolute URL; ends in a trailing slash when no logo is set.
          example: "https://bands.baphomet.club/images/band/logos/"
        image:
          type: string
          example: "https://bands.baphomet.club/images/band/images/amorphis.jpg"
        instagram: { type: string, nullable: true }
        spotify: { type: string, nullable: true }
        appleMusic: { type: string, nullable: true }
        bandcamp: { type: string, nullable: true }
        description:
          allOf: [{ $ref: "#/components/schemas/LocalizedText" }]
          description: Localized band description (German default + English).

    Stage:
      type: object
      description: Returned by `GET /api/festivals/{festivalSlug}/stages` and `.../stages/{stageSlug}`.
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: "Hauptbühne" }
        slug: { type: string, example: "hauptbuhne" }
        location: { type: string, nullable: true, example: "Dong, Neukirchen-Vluyn" }

    BandDetail:
      type: object
      description: Returned by `GET /api/festivals/{festivalSlug}/{bandSlug}`.
      properties:
        band:
          type: object
          properties:
            id: { type: integer, example: 8 }
            name: { type: string, example: "Amorphis" }
            genre:
              type: string
              nullable: true
            logo: { type: string, example: "https://bands.baphomet.club/images/band/logos/" }
            image: { type: string, example: "https://bands.baphomet.club/images/band/images/amorphis.jpg" }
            instagram: { type: string, nullable: true }
            spotify: { type: string, nullable: true }
            appleMusic: { type: string, nullable: true }
            bandcamp: { type: string, nullable: true }
            description:
              allOf: [{ $ref: "#/components/schemas/LocalizedText" }]
              description: Localized band description (German default + English).
        startTime:
          allOf: [{ $ref: "#/components/schemas/PhpDateTime" }]
          description: Present only when the band has a time slot at this festival.
        endTime:
          allOf: [{ $ref: "#/components/schemas/PhpDateTime" }]
          nullable: true
        stage:
          type: string
          description: Stage name. Present only when the band has a time slot.
          example: "Hauptbühne"

    TimeSlot:
      type: object
      description: Returned by `GET /api/festivals/{festivalSlug}/stages/{stageSlug}/timeslots`.
      properties:
        startTime: { $ref: "#/components/schemas/PhpDateTime" }
        endTime:
          allOf: [{ $ref: "#/components/schemas/PhpDateTime" }]
          nullable: true
        band: { type: string, example: "Aereum" }
        bandSlug: { type: string, example: "aereum" }
        bandId: { type: integer, example: 1 }
        stage: { type: string, example: "Hauptbühne" }
        stageSlug: { type: string, example: "hauptbuhne" }

    UpNext:
      type: object
      description: Returned by `GET .../upnext`. Here `startTime` is a preformatted string, not a PhpDateTime.
      properties:
        startTime:
          type: string
          description: Formatted as `H:i:s d.m.Y ` (trailing space included).
          example: "21:00:00 16.07.2026 "
        band: { type: string, example: "Amorphis" }
        stage: { type: string, example: "Hauptbühne" }

    LocalizedText:
      type: object
      description: |
        A piece of text in the API's supported languages. `de` (German) is the
        default and should be shown unless the client explicitly prefers English.
        Either field may be null when no copy exists for that language; clients
        fall back to the other language in that case.
      properties:
        de:
          type: string
          nullable: true
          description: German text (default language).
          example: "Melodic-Death- und Folk-Metal-Band aus Duisburg …"
        en:
          type: string
          nullable: true
          description: English text. Null when no English copy is available.
          example: "Melodic death and folk metal band from Duisburg, Germany …"
      required: [de, en]

    PhpDateTime:
      type: object
      description: |
        Symfony ObjectNormalizer serialization of a PHP `\DateTime`. Clients typically
        only need `timestamp` (Unix seconds, UTC). Other keys describe the timezone.
      properties:
        timestamp:
          type: integer
          description: Unix timestamp in seconds.
          example: 1784160000
        offset:
          type: integer
          example: 0
        microsecond:
          type: integer
          example: 0
        timezone:
          type: object
          properties:
            name: { type: string, example: "UTC" }
            transitions:
              type: array
              items: { type: object }
            location: { type: object }
      example:
        timestamp: 1784160000
        offset: 0
        microsecond: 0
        timezone:
          name: "UTC"